Doi badge

+++ Makefile
@@ -1,9 +1,4 @@
-VERSION=$(shell python -c "import quadpy; print(quadpy.__version__)")
-
-# Make sure we're on the master branch
-ifneq "$(shell git rev-parse --abbrev-ref HEAD)" "master"
-$(error Not on master branch)
-endif
+VERSION=$(shell python3 -c "import quadpy; print(quadpy.__version__)")
 
 default:
 	@echo "\"make publish\"?"
@@ -11,16 +6,19 @@ default:
 README.rst: README.md
 	cat README.md | sed 's_<img src="\([^"]*\)" width="\([^"]*\)">_![](\1){width="\2"}_g' > /tmp/README.md
 	pandoc /tmp/README.md -o README.rst
-	python setup.py check -r -s || exit 1
+	python3 setup.py check -r -s || exit 1
 
 # https://packaging.python.org/distributing/#id72
 upload: setup.py README.rst
+	# Make sure we're on the master branch
+	@if [ "$(shell git rev-parse --abbrev-ref HEAD)" != "master" ]; then exit 1; fi
 	rm -f dist/*
-	python setup.py bdist_wheel --universal
+	python3 setup.py bdist_wheel --universal
 	gpg --detach-sign -a dist/*
 	twine upload dist/*
 
 tag:
+	@if [ "$(shell git rev-parse --abbrev-ref HEAD)" != "master" ]; then exit 1; fi
 	@echo "Tagging v$(VERSION)..."
 	git tag v$(VERSION)
 	git push --tags
